fixed CollectionUtility.equalsCollection and improved performance
The equality check that is not forcing the same order of elements was
not correct. Collection.containsAll checked only one direction and did
not take the number of elements into account. So [a, a, b] was equal to
[a, b, b] or [a, b, c] was equal to [a, b, b].

Further, if the two collections were Lists and the order of elements was
not considered, the runtime complexity was O(n^2). Now it is O(n).

Change-Id: I9f4448a4ecf135e45d5d125dd2051d4ffb69716d
Reviewed-on: https://git.eclipse.org/r/75044
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>