Mobile: Make sure on screen keyboard can close
The focus manager prevents blur in certain circumstances (e.g. if a
button or an empty, non focusable area gets touched). This is bad on
touch devices because it prevents that the keyboard can close. This
leads to weird behaviour (e.g. on android if keyboard is closed using
back button, every touch opens it again because the previously focused
field still has the focus).