Mobile: Optimize Smart and DateField popup
The field on the touch popup didn't always show the same value as the
original field. The sync worked sometimes because the server sent the
display text back. This didn't happen in every case (e.g. if display
text was empty) -> make sure sync happens on gui correctly. Because the
popup was kept as variable, it would have been necessary to sync in both
directions. To avoid this the popup variable will only be set if the
popup is open -> makes it a lot easier to handle and is more consistent
with other widgets (e.g. menu)

The fix in the model adapter was necessary because not all properties
were in the list of modelProperties. The concrete problem was: If the
displayText was initially set it was correctly copied to the clone. If
not, the property was not in the list of modelProperties and therefore
not copied.