Fixed problem when adapted button is focused.
When a button was adapted by a ButtonAdapterMenu and that button was
configured to have initial focus an error occurred, because the button
was not rendered. To solve the problem we must delegate to the adapter
menu, which provides the DOM element to be focused, since the original
button is never rendered.

See ticket #174036.

Note: focus handling in complex scenarios (e.g. wizards) with multiple
inner forms and WrappedFormFields is still not working perfectly. There
are two open issues:
1. In WrappedFormField.js when _renderInnerForm is calls this.addField
it destroys the focus when addField calls jQuery append (DOM is changed)
2. When multiple inner forms are present, only one form should set the
initial focus. Currently more than one form sets the focus.