Fixed problem with wrong table selection and user-filters.
When a table with user-filters has been reloaded, the table-selection in
the UI was different from the UI-selection in the (Java) model. The
model added always an additional row to the selection because the state
of the user-filter was not correct because Table.js coalesced add and
remove events. In this case only the remove event was sent to the
UI-server but not the 'add'-event which calls setFilteredRowsFromUI.

Another error caused by this bug: because the model has a
multi-selection and a user wants to edit a row by clicking on a menu,
sometimes the wrong row was edited because the model simply takes the
first row of the multi-selection, the UI however shows only a single
selected row.

See ticket #175700

Conflicts:
	org.eclipse.scout.rt.ui.html.test/src/test/js/scout/table/TableSpec.js