Strong Content Security Policy
inline-javascript was only used in html files. These scripts have to be
externalized in a js file (good practive: with the same name as the html
file in the corresponding res folder).
The html file references this script file in the head section.

unsafe-eval is disabled by default, projects may override