Fixed problem with async loaded IMGs in scrollable table.
This fix adresses multiple problems:

1. when an IMG has been loaded, the table always scroll to the top
2. the row height was incorrectly set to the default row height, even
though the image was much larger
3. the width of the table (in a proposal chooser popup) has always
changed, because when an IMG has been loaded the layout has been
invalidated. For that I introduced a fixedWidth flag. So the table does
not change its width anymore, once the proposal chooser has set its
initial size.
4. scout.graphics.prefSize destroyed scroll positions because it changed
width and height of a $container. So we need to store and restore scroll
positions before and after we change the size.

See ticket #173870, #173428