Fixed UI error when form-validation fails for inner form fields.
The fix in activateProblemLocation() looks if the field is inside a
wrapped form field, and traverses further up the parent group boxes in
order to find a tab box. Doing so, the UI selects the tab box before it
gets the focus. This also means the tab box is rendered at that point
which was not the case before. Also added a safety-check in
FormField#focus().

See ticket #176022

Conflicts:
	org.eclipse.scout.rt.ui.html/src/main/js/scout/form/fields/FormField.js
	Didn't merge this, focus method doesn't exist in 5.2

Change-Id: Ia333b11827ddfeefa40ae8ca40e32211437e3b76