Mobile: Use column width as min width for auto resize tables
Currently the column width is used as weight for the autoResizeColumns
mechanism. To make sure the columns don't get too
small on small screens, this width is now also used as minimum /
preferred width for tables. It is not a hard minimum, the user can still
make the column smaller.