Disabled virtual scrolling in proposal-chooser table.
This prevents a problem with rows that have a bitmap image icon, which
triggers an invalidate layout when the image is loaded (asynchronously).
I tried many different solutions for that problem, but every other
solution caused another problem when virtual scrolling was on. The main
problem is, that the row.height is resetted when the width of the column
changes. But when we don't change the width of the table, the table is
probably too narrow and not all text / image is visible. So we decided
to simply deactivate virtual scrolling, because usually the lookup call
limits the rows returned by the table.

See ticket #173870, #173428