Table: Make table header menu responsive
Table header menu can currently not be used on small devices
(smartphones). The new compact mode moves the right filter column to the
bottom of the first column. The compact mode gets active if table menu
does not fit into the screen.

This made it necessary that the menu gets scrollable because it is
bigger than before. An additional body element had to be added to make
the whither work (whither is positioned outside the div which is not
possible anymore when it is scrollable).

The additional html components were necessary to make getPreferredSize
of the filter column work.