fix glasspane target for outlines
when opening a non-modal form with the desktop as display parent from
within a modal form that has the outline as display parent the newly
opened non-modal form should not be blocked. Equally other forms added
to the bench should also not be blocked if they do not share the same
display parent with the modal form.

Change-Id: I5427cc7aaddca2648f0f71a6c984fe0645bc0455
Signed-off-by: Michael Rudolf <mru@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/76715
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>