Increase time resolution in waitForBackgroundJobs()
When many background jobs finish but don't need to wake up the poller
(e.g. when receiving a large number of notifications from a service that
don't apply to the current user), the maximum delay could be exceeded.
This was caused by rounding errors because waitForBackgroundJobs() did
all calculations in units of seconds. When notification tokens occurred
multiple times per second, the elapsed duration was cut off to 0, which
caused the loop to wait for the same amount of time again, even tough
some time was elapsed since. This could lead to the total duration of
the polling request to become much larger than 60s, and the UI would
cancel the request after 75s.

The fix consists of two parts:
1. Use milliseconds instead of seconds internally (better resolution,
   no need to round and possibly lose precision)
2. Always measure elapsed time since start of method (t0) instead of
   adding individual loop durations.

176547

Change-Id: I46301abaa64d1205aa419e54292cc0224b33752b
Reviewed-on: https://git.eclipse.org/r/77235
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>