Added input stream reader that removes possibly available BOM
Unicode uses a byte order mark (BOM) for detecting the endianness of
the encoded stream. The character \ufeff represents the BOM, but only,
if it appears at the beginning of a stream. Within the text it
represents a ZERO WIDTH NON-BREAKING SPACE ZWNBS (it is strongly
recommended to use another character within the text for expressing the
same semantics, but ZWNBS is still supported).

The BomInputStreamReader transparently removes the BOM from unicode
streams. UTF-32 as well as other encodings is not affected by the
implementation.

Added comment to CsvHelper, that it should be used with a
BomInputStreamReader. Also made CsvHelper using its getters instead of
directly accessing its fields.

176035

Change-Id: Ib5cf1e7e0388941c45164fa4a6339230651aae2c
Reviewed-on: https://git.eclipse.org/r/77393
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>