Restore scroll position after setting the focus
When the FocusContext sets the focus to a field or a container, any
parent container that is scrollable might change its scroll position.
This is because the browser tries to scroll the focused field to the
visible range. Because this might produced unexpected "jumps", we
remember the old scroll positions and restore them after we have set the
focus.

Change-Id: Iaae416e271a5a054d18aea86a4fe5aef9f714bbd