Scout JS: Add NLS support
TextsLoader gathers all texts for the specified locale and builds a
texts.json.

This file is loaded by JavaScript on startup. The texts are stored
separately for each locale in a scout.Texts object. These text objects
are linked to support a parent lookup. The root text object is stored on
the session.

In the remote case there is still only one text object, stored on the
session as well -> Same mechanism for remote and local case may be used.