Prevent inserting text when result length would exceed "maxLength"
When the model wants to insert text into a string field, the UI should
check if the result would fit. Otherwise, the result would be truncated
at the end. Instead, a warning message is displayed as notification.

Ensure that PROP_MAX_LENGTH of StringField is never negative. 0 is a
valid value (although not very reasonable). Special logic in
getter/setter was removed.

177274

Change-Id: I5aff3f52be5f85886974091edce269aec4267f79