Model adapter split: Widget cloning
Compared to the previous clone mechanism, the new one does only clone
widgets, not adapters. And the clone is not registered in a registry
anymore. If something happens on a clone (e.g. click on a menu item) it
needs to be delegated to the original. Since the adapter is attached to
the original widget, it will be informed about that change and can sync
it to the server. If there is no adapter, everything will still work ->
Main advantage to before, it works without adapters.