Remove "validateRoot" from DesktopNavigation and SimpleTabBox
DesktopNavigation:
Setting "validateRoot=true" on the outline had no effect, as this
property should be set on the HTML component. Because the distance to
the desktop is not far, the line was removed completely (instead of
fixing it).

SimpleTabBox:
Not in all cases are simple tab boxes the validate root. Because the
distance to the desktop is not far, the line was removed completely
(instead of fixing it).

Change-Id: Iaa1824d7767cabe2bea110f07bf4b06af1357ab6