Fix JS error in FF when leaving page (new property "unloading")
When the application is open in FF and a new URL is entered or a
bookmark is clicked, a JS error occurs in Scout that is visible for a
short amount of time and is logged to the server. The reason for that
error is, that goOffline() is executed before _onWindowUnload() is
called where the "unload" flag is set.

To fix the error, a new "unloading" flag is already set in
_onWindowBeforeUnload().

Change-Id: I81608bf5282705c14544b515cb8c6447935462ff