Table: Outline shows the wrong nodes after resetting table
If a filter is active and the table organize menu is used to reset the
table (which removes the filters), the nodes in the outline are still
filtered.
Reason: Table rows are deleted and inserted again without executing the
tree filter -> Make sure tree filter is applied when new rows get
inserted.

177055