Fix position of tooltips and popups when $parent is not at (0,0)
When the parent of the tooltip and popup DIVs (usually the .scout div)
is not at position (0, 0) of the current window, the offset has to be
removed from the calculated positions. The reason is, that in most cases
the target position is retreived using the offset() method, which
returns an absolute position on the document, while the "top" and "left"
CSS properties of the tooltip/popup DIVs is relative to their parent.

Additional change:
All rendered Widgets now have the variable this.$parent set.

Change-Id: I08b2e2aa91a11e57af232213467566cc8131b883