Various layout fixes in MenuBar and MenuBox
MenuBar:
Adding the 'last' marker was broken because menus are no longer directly
added to the menubar $container, but to $left or $right. The algorithm
to find the last item was simplified (was complicated, because the right
aligned menus used to be stored in reverse order).

MenuBoxLayout:
Added functionality to add 'first' and 'last' CSS classes to the first
and the last visible menu item in the menu-box. The markers have to be
updated, too, when calculating the preferred size, because the classes
could imply different paddings/margins etc.

Additionally, the positioning of menu popups was fixed. The position
cannot be correctly set during the layout, but only after the entire
desktop was layouted. The LayoutValidator can now store functions that
are to be executed after the layout was validated.

DesktopToolBox:
The right margin of the last menu item was removed.

DesktopHeaderLayout:
Space should not be reserved for the tabbar's "ellipsis" menu when there
are not tabs.

menus:
Fixed the order of menus in the ellipsis menu.

Change-Id: I444f270d43b0241250e88462875fc6aa7bc766bf