Outline: Fix deferred glass pane rendering
Usecase: Outline is not in front and outlineContent is not rendered yet
(may happen if view is open and the page changed in the background so
that outlineContent.remove() is called but new outlineContent not
rendered yet). A form is opened with outline as displayParent. The form
won't be visible because outline is still in the background. As soon as
the outline comes to front, the form will be rendered.

An exception occurs because the outline content is not rendered and the
GlassPaneRenderer fails to render a glasspane (glassPaneTarget is
undefined -> $glassPane.window(true).popupWindow fails).

Fixed by using the DeferredGlasspaneTarget for the outline content. Also
consolidated and simplified the deferred handling.

176610