Model adapter split: Improve destroy behaviour
Since the widgets may now be destroyed, it is essential that they will
be destroyed eventually. And, the must not be destroyed if they are not
allowed to.

Use cases:
1. Table has menus. If the table gets new menus, the old ones need to be
destroyed, incl. their adapters.
2. View menu popup contains view menus. These view menus are maintained
by the desktop. When the popup closes, the menus must not be destroyed,
because they will be reused the next time the popup opens again.
3. The form is a shared widget and must only be destroyed when it is
closed. The servers sends a disposeAdapter event for this reason, until
then the form must not be destroyed, even if the parent is. Actually the
same as 2, beside the owner is set by the server.

OwnedAdapters is not necessary anymore, the destroy will happen at the
widget hierarchy. Hence the adapter does not destroy its children
anymore, the widget takes care of that. The adapter attaches a destroy
listener and destroys itself when the widget is being destroyed.