fixed scope propagation during registration and object initialization
The extension scope is augmented by each scout model object and all its
registered extensions as long as its initConfig method is running. This
is required for filtering extensions and for creating new instances of
them. Unfortunately, once an extension scope was established, it was
used for the complete initialization sequence that might create a client
session, a desktop, all available outlines with all their top-level
pages and search forms.

This change adds the possibility for scout model objects to request a
new extension scope which is used during the initialization (e.g. a form
is created using its own extension scope, but a form field reuses the
available scope). Whether or not a new scope is required is controlled
by a constructor parameter of ObjectExtensions.

Further, some model objects belong to others, but they are not
initialized together. Currently, this applies only to form and form
handler. A form handler is typically declared within a form class, but
it is not created by the form's initConfig method. Hence a form handler
extension that is declared within a form extension was never
instantiated because neither the form nor the form extension was on the
current extension scope.

The new InheritOuterExtensionScope annotation controls, whether the
outer extension's owner is used as filter for the annotated extension.
If the annotation's value is true or the annotation is missing at all,
the outer owner is used as filter. Otherwise the extension is registered
in the global scope. Currently only AbstractFormHandlerExtension is
annotated with @InheritOuterExtensionScope(false).

177873

Change-Id: Iec0a7906a7d3c61cf6d3085f44c445018b73a130
Reviewed-on: https://git.eclipse.org/r/79207
Tested-by: Hudson CI
Reviewed-by: Stephan Leicht Vogt <stephan.leichtvogt@bsi-software.com>