Widget: Links widgets on property change
If a widget is attached to another parent, setParent has to be called.
Until now this was mostly done manually in renderXY. With the new
concept setProperty takes pare of linking the widgets, so manual action
is only required if the widget does not use setProperty.

Also, if the parent is destroyed and is not the owner, the child widget
is now relinked to the owner so that the destroyed parent does not keep
a link to the child (improves garbage collection)