Model adapter split: Refactor setter and sync methods
Before the split, the model adapter set all the properties on a a
property change first or called the sync method, rendered all the
properties in a second step and finally fired a property change event so
that other widgets may react. 

With the new concept the model adapter just calls the setter of the
widget which is either setProperty or, if implemented, setXY. The setter
then either directly sets the property using _setProperty or calls the
sync method, if implemented. Afterwards the render method is called. The
main difference is that no property change is fired automatically. This
means if the sync method is implemented, the widget should use
_setProperty in order to fire a property change. 
Furthermore: 
- Return false has no effect anymore in the sync methods.
- setProperty may be used for regular properties
- The setter should not be called in the init method anymore, instead
use the sync method. Reason: the setter does an equals check at the
beginning. Because every property is written to the widget in init(), it
would always return true.