Default values: Correctly filter values in hierarchy
When different default values are defined in the hierarchy, only the
"nearest" default value has to be considered when checking if the actual
value is equals to a default value and can be ignored.

Example:
- "FormField" defines a default value "statusVisible=true"
- "Button" defines a default value "statusVisible=false"
- "Button" has "FormField" as object type parent
Previously, it was not possible to make the status visible for a button,
because the default value filter removed it in both cases. However, it
is wrong to filter the value when comparing it to "FormField"'s default
value, because "Button" already defines a default value.

177804

Change-Id: Ie6f06c1b24e6ee0d134aed9a96ad37940d78d03b