Model adapter split: Move goOffline/goOnline to adapter
Currently the widgets will get disabled when no connection is available,
because the widget need to ui server to work. For JS only apps this is
not true anymore, the widgets work without ui server and don't need to
be disabled.
Therefore to goOffline/goOnline methods have been moved to the adapter
which can disabled its widget if required. The adapter saves the old
enabled state and may therefore use the proper api method setEnabled()
instead of renderEnabled. This makes the parameter at renderEnabled
obsolete.

Also moved renderEnabled and renderVisible to Widget.js.

Also created postRenderActions in Widget.js to postpone the detach
action when it is executed while the widget is rendering. This was
necessary for the SimpleTabBox so that renderVisible may be called in
renderProperties.