Model adapter split: Remove old widgets on property change
Before the split, the old widgets were removed on a property change if
the property was an adapter property. This functionality has been
deleted assuming the widgets implement _removeXY anyway. It turned out
that this is still needed (e.g. by the Table for the table controls) and
therefore was added again.