Move Transaction API into platform module
Motivation:
Allows the usage of transaction even in client, and is a prerequisite
for MOM (Message Oriented Middleware).

Impact:
RunContext demarcates transaction boundary (former only
ServerRunContext). However, transaction scope of ServerRunContext is not
changed, meaning that it still starts a new transaction. However,
RunContext also starts a transaction, but only if not running in a
transaction yet.


Change-Id: I50495c39f6a00f170fef1e50631843faed5f58a6