ContextMenuPopup: Fix cleanup of menu adapter clones
Under some circumstances, unregisterAllAdapterClones was called twice,
which led to a JS error. To fix this, the menu adapter clones should be
unregistered recursively, when the ContextMenuPopup is removed. The
'remove' event on the menus is not sufficient, because sub-menus might
not be rendered, and so the 'remove' event is not triggered.

178000

Conflicts: [CP from 5.2.x]
	org.eclipse.scout.rt.ui.html/src/main/js/scout/menu/ContextMenuPopup.js

Change-Id: Iec88c43e490d30d9c3aa040030d89732b13a2cba