JsonOutline: Don't create adapters for filtered menus
JsonOutline filters the menus to only send the header menus.
Unfortunately, the filter only prevented the menu to be in the property
list but the adapter for the menu was created anyway. This led to a JS
exception called 'No adapter registered for ID' because no parent
created an adapter for the menu on js side because the outline did not
contain the menu in the property 'menus'. 

Solution: Made sure that no adapter is created if it is filtered ->
Refactored JsonContextMenu so that it does no longer extend from
AbstractJsonAdapter. It is now more or less a helper to handle the
structure changed event.