Make current RunContext accessible via 'RunContext.CURRENT' thread-local
- facilitates capturing of the current RunContext in shared code
- simplifies implementation of RunContext classes
- RunContexts.copyCurrent is based on RunContext.copy

Also, the creation of RunContext object was moved into RunContexts
classes. This simplifies maintenance of RunContext classes and follows
the pattern 'separation of concern'.

This change implies that copying the current context requires to run in
a RunContext (as given by its name). If not given and copying the
current context, this might be a programming error and should be
revisited.


Change-Id: Ic36e687fba516b4e4c639f1d8dcd86301150aaef