PopupWithHead: Don't fail on resize event if not rendered
Case: Context menu has a form menu as menu item. If form menu is
selected and the window resized and exception occured because $parentw
as null. Reason: Context
menu handles the resize event and closes the popup which removes the
form menu. Form menu detaches the resize listener but this has no effect
because the browser is already triggering the events.