Model adapter split: Cleanup addEventSupprt/addKeystrokeContextSupport
Because actually every widget needs event support is was added to
Widget.js.
KeystrokeSupport is not needed always, for this reason the widget had to
call addKeystrokeContextSupport. To make the code simpler, the add
method has been removed and the initialization moved to the existing
method initKeystrokeContext. If keyStrokeContextSupport is required, one
has to override _createKeyStrokeContext and create one instead of using
addKeyStrokeContextSupport. Furthermore the parameter has been removed. 

That means super has to be called in initKeyStrokeContext and
this.keyStrokeContext has to be used instead of the parameter.