Tooltip: Fix exception
Two exceptions, tooltip is in a groupbox of a tabbox

1. Second tab is selected, tooltip is being shown in the (hidden) first
tab: Exception because this.$parent was empty due to detached tab.
2. Second tab is selected, tooltip is already shown in the (hidden)
first tab. To parent of the tooltip gets destryoed incl. the tooltip ->
exception when switching back to first tab because it tries to render a
destryoed tooltip.

Furthermore: Every widget now triggers a render event