Widget: Child widget must not be removed if parent is not the owner
Usecase: Removing outline and hiding desktop navigation while default
detail form is open. Expected result is that default detail form stays
open, even if the owner (outline) is removed. Current parent of the
default detail form is the desktop bench which is still rendered.

If outline would be destroyed, default detail form will be destroyed and
removed as well.