ModelAdapter: Remove uniqueId
Having the attribute id filled for every widget (previously only for
every model adapter widget) is not required anymore. To have the id in
the DOM for debug purpose, it is added using scout.inspector, but only
if inspector is enabled. Whether inspector is enabled or not is stored
on the session and sent by the server using startup response.