Fixed bug in BrowserField JSON layer.
BinaryResourceUrlUtility.createDynamicAdapterResourceUrl returned the
same URL for different resources with the same filename (e.g.
'content.html') although the content and the fingerprint of the files
was different. Because the URL is the same, the UI didn't load the
binary resource, when a new binary resource with the same filename was
set on the BrowserField. To solve the issue I added the fingerprint to
the URL (separated by / slash). The fingerprint-part is removed in the
JSON layer and only the filename-part is passed to the BrowserField.
Thus the fingerprint is only needed to create a unique URL for the
resource.

Also fixed wrong variable name in ModelAdapter.js