Table: remove notifyServer
Also added filterCount check in deleteRows and deleteAllRows, otherwise
rowsFiltered event would be sent if all rows got deleted. UpdateRows
already uses the same check.

RowsSelected and rowsChecked filters don't check for the equality of the
rows due to performance reasons and because it should not be necessary.
The client doesn't select other rows while a rowsSelected or any other
event is being processed.

Also removed not used event columnActionsChanged.