Table: Improve page up/down scrolling
If the selected row is larger than the viewport and the only visible in
the viewport, nothing happens when pressing page up or down.
Also if the selected row is larger than the viewport and if there is
another one visible but not fully visible, an error occurs when pressing
page up or down.

178201