Preserve certain events in JsonResponse for new adapters
When a JSON adapter is created, events are usually not sent to the UI,
because the event content is already contained in the adapter data.
However, there are certain "action" events, that should be preserved and
executed in the post render phase of the widget:

  Desktop:
  - EVENT_OUTLINE_CONTENT_ACTIVATE
  - EVENT_ADD_NOTIFICATION
  - EVENT_REMOVE_NOTIFICATION
  - EVENT_OPEN_URI
  - EVENT_FORM_ACTIVATE

  Form:
  - (EVENT_REQUEST_FOCUS *)

  * This event is currently not marked as protected, because there
  is already a special property "initial focus" for that case. A TODO
  marker was added to change this in the future.)

  Table:
  - EVENT_REQUEST_FOCUS
  - EVENT_SCROLL_TO_SELECTION
  - EVENT_START_CELL_EDIT
  - EVENT_END_CELL_EDIT
  - EVENT_REQUEST_FOCUS_IN_CELL

  Tree:
  - EVENT_REQUEST_FOCUS
  - EVENT_SCROLL_TO_SELECTION

JsonEvent can be protected by calling event.protect(). In the widget,
the model action handler should check the this.rendered property and
schedule a "post render action" if the widget is not rendered.

For example, this allows adding a notification in the desktop's
execOpened() method.

177042
172649

Change-Id: I17d076b41b255223f023c69f3679ce4b2e9a0cbc