ClipboardField: Support text paste with disabled ClipboardAccess in IE
The ClipboardField had already a fallback method to get pasted image
content when the onPaste-event-data is empty (which will happens when
“Allow programmatic clipboard access” in IE is disabled for the security
zone of the scout application. An onPaste-event will still be fired, but
there is no way to get the clipboard content from the event-data). This
change adds the capabilities to deal with text content on the fallback
method.

177894

Change-Id: I9a76dbacc15b1cd55bc37e75c6da997d3ac86f99
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/80303
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>