Use fix modelId instead of model.objectType to lookup model json.
This solves a problem where a subclassed base-class wants to load a
model json. In that case we must not use the objectType of the subclass
(=virtual) but we still need the original objectType of the base-class.
Since JS does not know a class-name (as string), we must write down the
class-name in the code.