Table: Fix rowOrderChanged exception
rowOrderChanged event contains null entries -> the array of rowIds is
bigger than excepted -> JS throws exception "Row order may not be
updated because lengths of the arrays differ."

Reason: The null entries come from filtered rows which don't have a row
id. Due to the flag ignoreAccepted, isRowAccepted is not executed which
is wrong for this case.

Another issue: If the filter is added and removed in the same request,
the delete event will be generated anyway even though the filter has
already been removed.

Solution: The flag ignoreAccepted feels like a hack and and may not work
in every situation. To remove the hack and solve the issues, following
steps are taken:

1. Don't convert the filter_changed at the time of occurrence, convert
it before sending the response to the client. This solves the issue with
the add and remove filter in the same request.
2. Move the delete event to the beginning. This makes the flag
ignoreAccepted unnecessary.
3. Move the cleanup of events before an insert event from table event
buffer to json table. This cleanup is only necessary for the filter
conversion because row events may normally only be fired after the row
has been inserted -> Does not solve a new issue, just solves it in
another way (see JsonTableTest.testRemoveRowFilterAfterUpdates)

177932
178211