Implement ensure types
Properties may be sometimes called with a plain object or string, or
with an object which already has the correct type. The first case
happens if the property is sent by the server using json, it also
happens if the property is defined in the local json model files. The
second case may happen if the property is set programmatically.