Tree: Don't re-decorate node on selection change
Re-decorating the node on selection change is unnecessary, as only the
'group' class has to be re-assigned correctly.

This also fixes a bug on Firefox, when the tree field did not get the
selection when the user click directly on the text of a node (because
apparently setting the same text again breaks the click event's target).

174759

Change-Id: Icf0a5a24d5e2c5edadba30fe9c1e86f102145992
Reviewed-on: https://git.eclipse.org/r/80811
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>