Fix removal of views with displayParent=View in desktop bench layout
When a view is removed from the desktop bench, this might trigger the
removal of other forms (if they have the first form as display parent)
which then might result in the removal of the entire SimpleTabBox or the
BenchColumn (because they are now empty). Because the first "removeView"
method has not finished yet, this can lead to UI errors.

Only the "outer-most" removeView call should trigger the layout. To
achieve that, a counter "_removeViewInProgress" was introduced. Only
when the counter reaches 0 the layout may be revalidated.

Fixes the following cases:
1.
- View1 opens View2 with displayParent=View1
- View2 opens View3 with displayParent=View2
- View2 is activated
- View2 is closed (automatically closes View3)
2.
- View1 is a detail form of a page
- View1 opens View2 with displayParent=View1
- Another outline is activated (should detach View2 until the first
outline is activated again)

[CP from 6.0.x]

Change-Id: Ie9a1f8dfb2187a922191b5e34827d3e834707c8d