FormMenuPopup should update $head when formMenu changes properties
When the underlying menu changes one of its visible properties (e.g.
text, icon, cssClass, ...), the $head in the FormPopupMenu should be
changed accordingly. The menu previously propagated text changes to the
popup. This code was removed and replaced by a generic 'propertyChange'
event listener on the popup that is attached to the menu.

172810


Change-Id: Idad743a6d7999825925fd54799b51a54933d4789
Reviewed-on: https://git.eclipse.org/r/81229
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>