FormMenuPopup should update $head when formMenu changes properties
When the underlying menu changes one of its visible properties (e.g.
text, icon, cssClass, ...), the $head in the FormPopupMenu should be
changed accordingly. The menu previously propagated text changes to the
popup. This code was removed and replaced by a generic 'propertyChange'
event listener on the popup that is attached to the menu.

172810

[CP from 6.0.x]

Change-Id: Iad86b2636a0cca6491ec535f3b4ed05a2b6f4725