New method Platform.peek
Platform.get is auto-starting the Platform. However, in unit testing
it is desired to peek for the current implementation without the
Platform to start. This change adds a new method Platform.peek.
At the same time Platform.stop will stop the platform as before but will
not set it to null. @RunWithNewPlatform has an additional property to
define the platform used by the test.

Change-Id: Ifc962aa3fa4782db6dbe55db05c0719b8cf14637
Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/81328
Tested-by: Hudson CI