DialogLayout: Prevent unexpected shrinking of dialog bounds
When the contents of a dialog are changed (e.g. field gets invisible, or
another tab is selected in a tab box), the entire dialog is re-layouted
and the new preferred size is set. This can be quite unexpected and
annoying for the user, especially if he manually enlarged the dialog
(e.g. to see more data).

To improve the behavior, the dialog layout now ensures that the new
dialog size cannot be smaller than the old. The old "only resize if
height changes" logic was removed, because it did not work. This fix
works fine in most cases with only minor drawbacks:
- When first tab small and the user switches to a larger tab, the dialog
still "jumps". When he switches back to the first tab, the size stays
the same (causing empty space or streched fields on the first tab).
- When a large portion of the dialog gets invisible, the dialog might
feel "too big" (empty space / streched fields).
- When the user manually shrinks the dialog and changes a tab, the
dialog might "jump" back to the bigger preferred size.

Generally, the new behavior should be much more user friendly. We
recommend that large and/or highly dynamic dialogs are realized as views
(which don't have that problem).

178385

Change-Id: I7ffb3586ef00dc3b4ecf1acb9d23fad3e9d7364e