Tree: Remove notifyServer
Also fixed checkNodes: AutoCheckChildren did not work correctly for more
than 2 levels, children on the 3d level were not marked.

Also fixed Compact Tree: Compact tree selects the first section node if
the title node is selected. This selection change was not propagated to
the server because every selection event was filtered.