Tree: Improve autoCheckChildNodes
1. If a node is disabled, the state does not get changed if
enabledNodesOnly is set to true. If this node has enabled child nodes,
these nodes would not get the correct state if the parent of the
disabled node is marked -> It is not correct to read the state of the
parent, the state has to be passed along.
2. execAutoCheckChildNodes does not consider enabledNodesOnly -> Pass it
along and call checkNodes to make sure the enabled check is performed
for every node