OpenUriAction "open" should not replace current location (except on iOS)
In Firefox, all ajax requests are killed when window.location.href is
set, even if the target is not a website (e.g. mailto: link or a
download). Therefore, the action "open" should use a hidden iframe. This
works on most platforms. On iOS, the URI must be set to
window.location.href (same "if" as with type "download").

[CP From 6.0.x]

Change-Id: Ie4488b16de54a5950a322ac873da17f44ead64f6