Codes: Improve NLS handling
Codes now accept either the text or a texts object. If a text is
provided, code.text() returns this text (or the resolved text if the
text is a text key reference). Otherwise the texts in the text object
are registered in the global text maps (scout.texts) with a generated
text key. The big advantage is that the existing parent / default lookup
functionality may be reused (see scout.texts for details).