ContextMenu: Show ellipsis if text is truncated
On small screens, context menu should truncate text if it is too long.
Icon and submenu icon are always visible, only the text is shortened. We
do not make the submenu icon invisible by purpose: It would be confusing
if one menu shows the icon and one not if both have sub menus.

178296